Development of control interface for keyboard
