+++ vehicle_user_control/README.md
@@ -1,0 +1,14 @@
+# Vehicle User Control 
+
+ROS Package for VantTec user interfaces for diverse control methods.
+
+Currently supports:
+- Keyboard
+
+*Dependencies*
+- SDL
+
+```
+sudo apt-get install libsdl1.2-dev
+sudo apt-get install libsdl1.2debian
+```
